carnot-attest/pyproject.toml
[project]
name='carnot-attest'
version='0.2.0'
dependencies=['pynacl','jinja2']
[project.scripts]
carnot-attest='carnot_attest.__main__:main'
